+++ src/components/header/AppHeader.tsx
@@ -2,17 +2,15 @@ import React from 'react';
 import headerStyles from './header.module.css';
 import { Logo, ProfileIcon, BurgerIcon, ListIcon } from '@ya.praktikum/react-developer-burger-ui-components'
 import { Link, useLocation, useNavigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import { rSignInSelector } from '../../services/redux/selectors/selectorsLogin';
 
 const AppHeader = () => {
 
   const navigate = useNavigate();
-  const { statusSign } = useSelector(rSignInSelector);
 
-  const onClickLogo = () => { if (statusSign) navigate('/profile') }
-  const onClickConstr = () => { if (statusSign) navigate('/') }
+  const onClickLogo = () => { navigate('/profile') }
+  const onClickConstr = () => { navigate('/') }
   
+
   const location = useLocation();
 
   return (
