+++ src/components/ingrDetail/ingr-detail.module.css
@@ -1,41 +1,11 @@
-.wrapper {
-    position: fixed;
-    width: 720px;
-    height: 539px;
-    margin-left: auto;
-    margin-right: auto;
-    left: 0;
-    right: 0;
-    text-align: center;
-    top: 15%;
-    z-index: 3;
-}
-
-.card {
-    box-sizing: border-box;
-    position: absolute;
-    left: 0%;
-    right: 0%;
-    top: 0%;
-    bottom: 0%;
-    background: #1C1C21;
-    box-shadow: 0px 24px 32px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 1px rgba(0, 0, 0, 0.04);
-    border-radius: 40px;
-}
-
 .cardNoModal {
-    box-sizing: border-box;
-    position: absolute;
-    left: 0%;
-    right: 0%;
-    top: 0%;
-    bottom: 0%;
+    display: flex;
+    justify-content: space-around;
 }
-
 .header {
     display: flex;
     flex-direction: row;
-    justify-content: flex-end;
+    justify-content: center;
     align-items: center;
     padding: 0px;
     gap: 36px;
@@ -43,7 +13,7 @@
     height: 64px;
     left: 40px;
     right: 40px;
-    top: 40px;
+    top: 120px;
 }
 
 .headText {
@@ -56,6 +26,7 @@
     line-height: 40px;
     display: flex;
     align-items: center;
+    justify-content: center;
 }
 
 .headTextNoModal {
@@ -79,7 +50,14 @@
     width: 520px;
     height: 30px;
     left: 100px;
-    top: 361px;
+    top: 481px;
+}
+
+.frameNoModal {
+    position: absolute;
+    width: 520px;
+    height: 30px;
+    top: 481px;
 }
 
 
@@ -107,7 +85,19 @@
     width: 516px;
     height: 56px;
     left: 102px;
-    top: 423px;
+    top: 543px;
+}
+
+.nutritionNoModul {
+    display: flex;
+    flex-direction: row;
+    align-items: flex-start;
+    padding: 0px;
+    gap: 20px;
+    position: absolute;
+    width: 516px;
+    height: 56px;
+    top: 543px;
 }
 
 .values1 {
@@ -183,5 +173,11 @@
     width: 520px;
     height: 240px;
     left: 100px;
-    top: 105px;
+    top: 225px;
+}
+.imgNoModal {
+    position: absolute;
+    width: 520px;
+    height: 240px;
+    top: 225px;
 }
