--- src/components/ingrModal/IngrModal.tsx
@@ -1,112 +1,0 @@
-import React, { useCallback, useEffect, FC } from 'react'
-import ReactDOM from 'react-dom'
-import moduleStyles from './ingrModal.module.css';
-import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components'
-import ModalOverlay from '../modal-overlay/ModalOverlay';
-import { useDispatch, useSelector } from "react-redux";
-import { dataSelector, idSelector, showIngrSelector } from "../../services/redux/selectors/selectorsIngr";
-import { actionIngr } from "../../services/redux/actionCreators/actionIngr";
-import { useNavigate } from 'react-router-dom';
-import { fetchData } from '../../services/redux/thunks/thunkIngr';
-import { getCookie } from '../../utils/func-cooke';
-import toast from 'react-hot-toast';
-import { T_Id } from '../../utils/types';
-
-
-const modalDiv = document.getElementById("modals")!
-
-type TIngrModalProps = {
-    ipProps?: string;
-}
-
-const IngrModal: FC<TIngrModalProps> = ({ ipProps }) => {
-
-    const dispatch = useDispatch() as any
-    const navigate = useNavigate();
-
-    const isShow = useSelector(showIngrSelector)
-    const data = useSelector(dataSelector)
-    const id = useSelector(idSelector)
-
-    useEffect(() => {
-        const accessToken = getCookie('accessToken')
-        if (!accessToken) {
-            //toastError('Только авторизованные пользователи могут просматривать ингредиенты')
-            navigate('/login', { replace: true })
-        }
-        if (data && accessToken) {
-            dispatch(fetchData())
-        }
-    }, []);
-
-
-    const dataModul = data.filter((item: T_Id) => item._id === id || ipProps )
-
-    const close = useCallback(() => {
-        dispatch(actionIngr.setShowIngr(false))
-        navigate(-1);
-    }, [isShow, navigate, dispatch])
-
-    useEffect(() => {
-        const modalDiv = document.getElementById("modalIngr")
-        const ModalOverlay = (e: MouseEvent) => { e.target === modalDiv && close() }
-        document.addEventListener("click", ModalOverlay)
-        
-        const esc = (e: KeyboardEvent) => { e.key === "Escape" && isShow && close() }
-        document.addEventListener("keydown", esc)
-
-        return () => {
-            document.removeEventListener("click", ModalOverlay)
-            document.removeEventListener("keydown", esc)
-        }
-    }, []);
-
-    return ReactDOM.createPortal(
-        <>
-            {dataModul[0] &&
-                <>
-                    <div className={moduleStyles.wrapper}>
-                        <div className={`${ id ? moduleStyles.card : moduleStyles.cardNoModal}`}>
-                        
-                            <div className={moduleStyles.header}>
-                                <span className={`${ id ? moduleStyles.headText : moduleStyles.headTextNoModal}`}>Детали ингредиента</span>
-                                {id &&
-                                <span className={moduleStyles.headClose} onClick={close}> <CloseIcon type="primary" /></span>
-                                }
-                            </div>
-
-                            <img src={dataModul[0].image_large} className={moduleStyles.img} alt="изображение ингридиента" />
-
-                            <p className={moduleStyles.frame}>
-                                <span className={moduleStyles.frameSpan}>{dataModul[0].name}</span>
-                            </p>
-
-                            <section className={moduleStyles.nutrition}>
-                                <div className={moduleStyles.values1}>
-                                    <span className={moduleStyles.kkal}>Калории, ккал</span>
-                                    <span className={moduleStyles.valname}>{dataModul[0].calories}</span>
-                                </div>
-                                <div className={moduleStyles.values2}>
-                                    <span className={moduleStyles.kkal}>Белки, г</span>
-                                    <span className={moduleStyles.valname}>{dataModul[0].proteins}</span>
-                                </div>
-                                <div className={moduleStyles.values3}>
-                                    <span className={moduleStyles.kkal}>Жиры, г</span>
-                                    <span className={moduleStyles.valname}>{dataModul[0].fat}</span>
-                                </div>
-                                <div className={moduleStyles.values4}>
-                                    <span className={moduleStyles.kkal}>Углеводы, г</span>
-                                    <span className={moduleStyles.valname}>{dataModul[0].carbohydrates}</span>
-                                </div>
-                            </section>
-                        </div>
-                    </div>
-                </>
-            }
-            <ModalOverlay id="modalIngr" />
-        </>,
-        modalDiv
-    )
-}
-
-export default React.memo(IngrModal);
