+++ src/components/modal/modal.module.css
@@ -7,7 +7,7 @@
     left: 0;
     right: 0;
     text-align: center;
-    top: 20%;
+    top: 10%;
     z-index: 3;
 }
 
@@ -29,143 +29,5 @@
     margin-top: 60px;
     margin-right: 40px;
     cursor: pointer;
-}
-
-
-.done {
-    position: absolute;
-    width: 120px;
-    height: 120px;
-    left: 300px;
-    top: 362px;
-}
-
-
-.ico {
-    position: absolute;
-    left: 50%;
-    top: 45%;
-    transform: translateX(-50%);
-}
-
-.vector1 {
-    position: absolute;
-    left: 17.76%;
-    right: 18.69%;
-    top: 17.65%;
-    bottom: 18.3%;
-    background: radial-gradient(50% 50% at 50% 50%, rgba(128, 26, 179, 0) 0%, rgba(76, 76, 255, 0.125) 100%) /* warning: gradient uses a rotation that is not supported by CSS and may not behave as expected */;
-    border-radius: 40px;
-}
-
-.vector2 {
-    position: absolute;
-    left: 0%;
-    right: 0%;
-    top: 0.98%;
-    bottom: 0.98%;
-    background: radial-gradient(50% 50% at 50% 50%, rgba(128, 26, 179, 0) 0%, rgba(76, 76, 255, 0.25) 100%) /* warning: gradient uses a rotation that is not supported by CSS and may not behave as expected */;
-    border-radius: 40px;
-}
-
-.vector3 {
-    position: absolute;
-    left: 3.74%;
-    right: 4.67%;
-    top: 0%;
-    bottom: 0%;
-    background: radial-gradient(50% 50% at 50% 50%, rgba(128, 26, 179, 0) 0%, rgba(76, 76, 255, 0.25) 100%) /* warning: gradient uses a rotation that is not supported by CSS and may not behave as expected */;
-    border-radius: 40px;
-}
-
-.redy {
-    position: absolute;
-    width: 520px;
-    height: 24px;
-    left: 100px;
-    top: 574px;
-    background: #1C1C21;
-}
-
-.redySpan {
-    position: absolute;
-    height: 24px;
-    left: 0px;
-    right: 0px;
-    top: 0px;
-    font-family: 'Jet Brains Mono';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    text-align: center;
-}
-
-.go {
-    position: absolute;
-    width: 520px;
-    height: 24px;
-    left: 100px;
-    top: 542px;
-    background: #1C1C21;
-}
-.goSpan {
-    position: absolute;
-    height: 24px;
-    left: 0px;
-    right: 0px;
-    top: 0px;
-    font-family: 'Jet Brains Mono';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    text-align: center;
-}
-
-.id {
-    position: absolute;
-    width: 520px;
-    height: 30px;
-    left: 100px;
-    top: 272px;
-    background: #1C1C21;
-}
-.idSpan {
-    position: absolute;
-    height: 30px;
-    left: 0px;
-    right: 0px;
-    top: 0px;
-    font-family: 'Jet Brains Mono';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 30px;
-    text-align: center;
-}
-
-.number {
-    position: absolute;
-    width: 520px;
-    height: 120px;
-    left: 100px;
-    top: 120px;
-    background: #1C1C21;
-}
-.numberSpan {
-    position: absolute;
-    height: 120px;
-    left: 0px;
-    right: 0px;
-    top: 0px;
-    font-family: 'Iceland';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 144px;
-    line-height: 120px;
-    text-align: center;
-    color: #F2F2F3;
-    text-shadow: 0px 0px 16px rgba(51, 51, 255, 0.25), 0px 0px 8px rgba(51, 51, 255, 0.25), 0px 4px 32px rgba(51, 51, 255, 0.5);
 }
