+++ src/services/redux/thunks/thunkLogout.js
@@ -25,7 +25,7 @@ export const fetchLogout = () => {
                 dispatch(actionAutReg.setStatus(false))
                 dispatch(actionConstr.clearConstr())
                 dispatch(actionForgResPas.setInitialState())
-                dispatch(actionIngr.setInitialState())
+                //dispatch(actionIngr.setInitialState())
                 dispatch(actionOrder.setInitialState())
                 dispatch(actionSignIn.clSignIn())
                 dispatch(actionOrder.setInitialState())
@@ -46,7 +46,7 @@ export const fetchLogout = () => {
                     dispatch(actionAutReg.setStatus(false))
                     dispatch(actionConstr.clearConstr())
                     dispatch(actionForgResPas.setInitialState())
-                    dispatch(actionIngr.setInitialState())
+                    //dispatch(actionIngr.setInitialState())
                     dispatch(actionOrder.setInitialState())
                     dispatch(actionSignIn.clSignIn())
                     dispatch(actionOrder.setInitialState())
